src/pa3_recognizer/recognition/requirements.txt
requests
opencv-python
ipython
PyHamcrest